perf: Improvement in the JS for overriding controls from core WP.
