+++ inc/customizer/extend-custom-controls/src/index.js
@@ -1,4 +1,4 @@
-import { BaseControl } from './base/control';
+import { coreControl } from './core/control';
 import { headingControl } from './heading/control';
 import { hiddenControl } from './hidden/control';
 import { descriptionControl } from './description/control';
